+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.1.1-dev
 
-* Support formatting enums.
+* Support formatting enums (#120).
+* Handle Windows line endings in multiline strings (#126).